Moved Async implementation to Internals module #review
* Add tests for Flat tree
* More comments suggested from review
